requirements.txt
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
numpy>=1.26.0
rich>=13.7.0      # pretty terminal output
typer>=0.12.0     # CLI for agent scripts
pydantic>=2.7.0   # data models
feedparser>=6.0.0 # RSS feed parsing for Research agent
